LRSD-8774 Refactor: Remove unnused and scope components
