+++ workspaces/liferay-learn-workspace/client-extensions/liferay-learn-custom-elements/src/components/Course/NavigationMenu.js
