--- workspaces/liferay-learn-workspace/client-extensions/liferay-learn-custom-elements/src/components/Duration.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Duration = () => {
-	return <div>Duration</div>;
-};
-
-export default Duration;
