--- workspaces/liferay-learn-workspace/client-extensions/liferay-learn-custom-elements/src/components/Enroll.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Enroll = () => {
-	return <div>Enroll</div>;
-};
-
-export default Enroll;
