--- workspaces/liferay-learn-workspace/client-extensions/liferay-learn-custom-elements/src/components/PageLoading.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const PageLoading = () => {
-	return <div>PageLoading</div>;
-};
-
-export default PageLoading;
