--- workspaces/liferay-learn-workspace/client-extensions/liferay-learn-custom-elements/src/components/SignIn.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const SignIn = () => {
-	return <div>SignIn</div>;
-};
-
-export default SignIn;
