--- workspaces/liferay-learn-workspace/client-extensions/liferay-learn-custom-elements/src/components/UserProgressReport.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const UserProgressReport = () => {
-	return <div>UserProgressReport</div>;
-};
-
-export default UserProgressReport;
